Suppress aegir eslint warnings
Most of the warnings could be eliminated, but it would be good
to discuss a strategy on a case-by-case basis. For example,
we do a lot of console logging for errors, but those should
probably be passed up so that the UI can handle them.

I propose that we suppress the warnings for now, and then file
some github issues so we don't suppress the warnings forever ...
eg. "remove console.log statements in collaboration/gossip", etc.