When creating batches, make sure all deltas are included
In high concurrency situations, often times delta will arrive
that depend on other deltas.

Right now, when we bundle deltas together into batches to send to
other peers, we are creating a clock for the batch that claims to
include all the deltas from a previous clock + an author clock, but
it is omitting any deltas that are dependencies of the ones that
are selected. As a result, when other peers receive the batch, they
will advance their vector clock ... but we currently have a problem,
because if they don't already have the dependencies missing from
the batch, they will never attempt to sync them.

This patch modifies the batch algorithm to check the previous clock
for the batch ... if new dependencies are added, when we compare it
to the 'since' clock that we are searching against, it will have
more content - so we restart the search for deltas with the wider
criteria so that we catch all the deltas.

There probably should be an additional check at the end of the
process to ensure that the author clocks add up, in case some of
the dependency deltas were trimmed and were not available.

A different approach would be to just abort creating the batch when
new dependencies are discovered (at the cost of creating more
work at the replication level).

In some situations, there might be several deltas stored that might
match the vector clock range used in constructing a batch (eg. the
original deltas, plus any intermediate batches). Right now, the
search algorithm just picks the first one that "is interesting".
I think that is okay, as the first one in the array is probably
going to be the smaller original delta.

In order to "rewind" the search when we need to include more
dependencies, I decided to throw an exception ... the logic was
complicated because of the way the batches array is constructed
due to sub-collaboration support. It could probably be rewritten
to be more efficient.

Also, I changed vectorclock.minimum so it doesn't default to zero
as the minimum for the clocks... eg. before, if one clock was '2',
and the other was '3', the minimum would be zero. For the algorithm
here to work, it needs to be '2' ... which is consistent to what
you'd expect with other minimum implementations, such as Math.min().
I think the vectorclock.minimum() function is only used by the
batching code. Also, I modified the spot where the vector clock is
first initialized so we explicitly store a '0' clock when we first
push data.

This seems to fix the main problem I was seeing with the
collaboration-random test not finishing. However, I am still seeing
it happen, but less often, so I will need to investigate some more.

There definitely needs to be some more test cases for how we want
deltas and clocks to combine into batches.